[C10D] Decouple PGNCCL desync from dbg dump
Add TORCH_NCCL_DUMP_DEBUG_INFO env to control dumping independently
of desync debug feature.

Currently default to disabled (so no behavior change by default),
but plan to default this to true after validation.

Keeps the 'sleep for 30 sec' that used to be after desync debug where it
was.

Fixes #114433

ghstack-source-id: f3ef7f99f9e9876ff0768fca877349908f3acba9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/114614